feat(redmine 1274413): extracted logic from ActionBar into new ActionRowOverflow
