+++ .changeset/strange-cooks-eat.md
@@ -1,0 +1,6 @@
+---
+'@smile/react-front-kit': minor
+'storybook-pages': minor
+---
+
+Extracted logic from ActionBar into new ActionRowOverflow, moved a type and some mocks
