+++ packages/react-front-kit/src/Components/ActionRowOverflow/ActionRowOverflow.stories.tsx
@@ -1,30 +1,27 @@
-import type { IActionRowOverflowAction } from '../ActionRowOverflow/ActionRowOverflow';
+import type { IActionRowOverflowAction } from './ActionRowOverflow';
 import type { Meta, StoryObj } from '@storybook/react';
 
+import { ActionRowOverflow as Cmp } from './ActionRowOverflow';
 import {
   actionRowOverflowActionsMock,
   actionRowOverflowSelectedMock,
-} from '../ActionRowOverflow/ActionRowOverflow.mock';
-
-import { ActionBar as Cmp } from './ActionBar';
-import { actionBarLabelMock } from './ActionBar.mock';
+} from './ActionRowOverflow.mock';
 
 const meta = {
   component: Cmp,
   tags: ['autodocs'],
-  title: '3-custom/Components/ActionBar',
+  title: '3-custom/Components/ActionRowOverflow',
 } satisfies Meta<typeof Cmp>;
 
 export default meta;
 type IStory = StoryObj<typeof meta>;
 
-export const ActionBar: IStory = {
+export const ActionRowOverflow: IStory = {
   args: {
     actions: actionRowOverflowActionsMock as IActionRowOverflowAction<
       Record<string, unknown>
     >[],
     rowActionNumber: 2,
     selectedElements: actionRowOverflowSelectedMock,
-    selectedElementsLabel: actionBarLabelMock,
   },
 };
