+++ packages/react-front-kit/src/index.tsx
@@ -2,13 +2,13 @@
 'use client';
 
 // component exports
-export type { IPreviewProps } from './Components/Preview/Preview';
-export { Preview } from './Components/Preview/Preview';
 export { ActionBar } from './Components/ActionBar/ActionBar';
+export type { IActionBarProps } from './Components/ActionBar/ActionBar';
+export { ActionRowOverflow } from './Components/ActionRowOverflow/ActionRowOverflow';
 export type {
-  IActionBarAction,
-  IActionBarProps,
-} from './Components/ActionBar/ActionBar';
+  IActionRowOverflowProps,
+  IActionRowOverflowAction,
+} from './Components/ActionRowOverflow/ActionRowOverflow';
 export { BitConverter } from './Components/BitConverter/BitConverter';
 export type { IBitConverterProps } from './Components/BitConverter/BitConverter';
 export { Breadcrumbs } from './Components/Breadcrumbs/Breadcrumbs';
@@ -50,12 +50,24 @@ export { Motif } from './Components/InfoCard/Motif';
 export type { IMotifProps } from './Components/InfoCard/Motif';
 export { Pagination } from './Components/Pagination/Pagination';
 export type { IPaginationProps } from './Components/Pagination/Pagination';
+export { Preview } from './Components/Preview/Preview';
+export type { IPreviewProps } from './Components/Preview/Preview';
 export { ResponsiveTabs } from './Components/ResponsiveTabs/ResponsiveTabs';
 export type { IResponsiveTabs } from './Components/ResponsiveTabs/ResponsiveTabs';
+export { SearchableCheckboxList } from './Components/SearchableCheckboxList/SearchableCheckboxList';
+export type {
+  ISearchableCheckboxListProps,
+  ICheckbox,
+} from './Components/SearchableCheckboxList/SearchableCheckboxList';
 export { SearchBar } from './Components/SearchBar/SearchBar';
 export type { ISearchBarProps } from './Components/SearchBar/SearchBar';
 export { SelectableList } from './Components/SelectableList/SelectableList';
 export type { ISelectableListProps } from './Components/SelectableList/SelectableList';
+export { SidebarFilters } from './Components/SidebarFilters/SidebarFilters';
+export type {
+  ISidebarFiltersProps,
+  ISidebarFilter,
+} from './Components/SidebarFilters/SidebarFilters';
 export {
   getRecursiveMenu,
   SidebarMenu,
@@ -64,16 +76,6 @@ export type {
   IMenuItem,
   ISidebarMenuProps,
 } from './Components/SidebarMenu/SidebarMenu';
-export { SearchableCheckboxList } from './Components/SearchableCheckboxList/SearchableCheckboxList';
-export type {
-  ICheckbox,
-  ISearchableCheckboxListProps,
-} from './Components/SearchableCheckboxList/SearchableCheckboxList';
-export { SidebarFilters } from './Components/SidebarFilters/SidebarFilters';
-export type {
-  ISidebarFiltersProps,
-  ISidebarFilter,
-} from './Components/SidebarFilters/SidebarFilters';
 export { SwitchableView } from './Components/SwitchableView/SwitchableView';
 export type {
   IDataView,
